Title: 사각형 채우기 3, Time: 89ms, Memory: 9MB, Status: Wrong Answer - Codetree

+++ 250809/사각형 채우기 3/rectangle-fill-3.java
@@ -20,7 +20,7 @@ public class Main {
         n = Integer.parseInt(st.nextToken());
 
 
-        int [] dp = new int[n+1];
+        long [] dp = new long[n+1];
       
         dp[1] = 2;
